:sparkles: 대체 파일 제거 및 유사 컴포넌트 병합
