+++ src/components/form-control/ToggleInput.jsx
@@ -1,7 +1,41 @@
 import styles from './ToggleInput.module.css';
 
 import { Controller } from 'react-hook-form';
-import ToggleSwitch from '@/components/ToggleSwitch.jsx';
+import { useId, useState } from 'react';
+
+function ToggleSwitch({ checked, value: valueProp, focusDisabled, onChange, ...props }) {
+  const isControlled = typeof checked !== 'undefined' || typeof valueProp !== 'undefined';
+  const isOn = isControlled ? !!(typeof checked !== 'undefined' ? checked : valueProp) : false;
+  const [focus, setFocus] = useState(false);
+  const id = useId();
+
+  return (
+    <label
+      className={`${styles.switch} ${styles.form}`}
+      htmlFor={id}
+      title={isOn ? '켜짐' : '꺼짐'}
+    >
+      <input
+        id={id}
+        type="checkbox"
+        checked={isOn}
+        onChange={(e) => onChange?.(e.target.checked)}
+        onFocus={() => setFocus(true)}
+        onBlur={() => setFocus(false)}
+        {...props}
+      />
+      <span
+        className={`${styles.slider} ${isOn ? styles.checked : ''} ${focusDisabled || !focus ? '' : styles.focused}`}
+        aria-hidden="true"
+      />
+      <span
+        className={`${styles.thumb}`}
+        style={{ transform: isOn ? 'translateX(22px)' : 'translateX(0)' }}
+        aria-hidden="true"
+      />
+    </label>
+  );
+}
 
 export default function ToggleInput({ label, name, control }) {
   return (
